timeout is an int
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1519103 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/coyote/ajp/AjpNioProcessor.java
@@ -90,7 +90,7 @@ public class AjpNioProcessor extends AbstractAjpProcessor<NioChannel> {
         // Setting up the socket
         this.socketWrapper = socket;
 
-        long soTimeout = endpoint.getSoTimeout();
+        int soTimeout = endpoint.getSoTimeout();
         boolean cping = false;
 
         // Error flag
